fixed some value conversion issues

+++ PNChart/PNLineChart.m
@@ -139,7 +139,7 @@
     UITouch *touch = [touches anyObject];
     CGPoint touchPoint = [touch locationInView:self];
 
-    for (int p = _pathPoints.count - 1; p >= 0; p--) {
+    for (NSInteger p = _pathPoints.count - 1; p >= 0; p--) {
         NSArray *linePointsArray = _pathPoints[p];
 
         for (int i = 0; i < linePointsArray.count - 1; i += 1) {
@@ -172,7 +172,7 @@
     UITouch *touch = [touches anyObject];
     CGPoint touchPoint = [touch locationInView:self];
 
-    for (int p = _pathPoints.count - 1; p >= 0; p--) {
+    for (NSInteger p = _pathPoints.count - 1; p >= 0; p--) {
         NSArray *linePointsArray = _pathPoints[p];
 
         for (int i = 0; i < linePointsArray.count - 1; i += 1) {
